수정: API 서비스 수정/추가 및 헬퍼 파일 삭제
