+++ app/services/haravan_product/get_haravan_products.rb
@@ -3,22 +3,15 @@ module HaravanProduct
     include HaravanApiHelper
     include HaravanApiHelper::HaravanProduct
 
-    def initialize
-      super
-    end
-
-    def new(*args)
-    end
-
+    # == 상품 불러오기 및 저장
     def save_haravan_products(from, to)
       ActiveRecord::Base.transaction do
         begin
           haravan_products = get_product_by_period(from, to)
           haravan_products.each do |haravan_product|
-            _parse_data(haravan_product)
-            _save_product_page
-            _save_product_options
-            true
+            extract_data(haravan_product)
+            save_product
+            save_product_options
           end
         rescue => e
           ap e
@@ -28,41 +21,52 @@ module HaravanProduct
 
     end
 
-
     private
 
+    @id
     @title
     @variants
     @brand
     @product
 
+    # == 데이터 추출
     # parse haravan raw data to make the products and product options
-    def _parse_data(haravan_product)
+    def extract_data(haravan_product)
+      @id = haravan_product["id"]
       @title = haravan_product["title"]
       @variants = haravan_product["variants"]
-      @brand = Brand.where('name LIKE ?', "%#{haravan_product["vendor"]}%").first
+      @brand = Brand.where("JSON_EXTRACT(name, '$.vi') LIKE ?", "\"#{haravan_product["vendor"]}\"").first
     end
 
+    # == 상품페이지를 저장
     # Product page saver
-    def _save_product_page
-      product_data = _make_product_data
-      @product = Product.where('title LIKE ?', "%#{@title}%").first_or_initialize
-      @product.assign_attributes(product_data)
-      @product.assign_attributes(title: product_data["title"].to_h)
+    def save_product
+      @product = Product.find_or_initialize_by(haravan_id: @id)
+      @product.assign_attributes(brand_id: @brand.id,
+                                 running_status: 'pending',
+                                 title: {'vn': @title, 'en': @title, 'ko': @title})
+      # @product.assign_attributes(title: product_data["title"].to_h)
       @product.save!
     end
 
+    # == 상품 옵션을 저장
     # Save Product Options that reveal on product page.
-    def _save_product_options
+    def save_product_options
       product_option_group = @product.option_groups.first_or_create
       channel_id = Channel.find_by_name('haravan').id
       @variants.each do |option|
-        product_option_params = _make_product_option_data(option, channel_id)
-        product_option = product_option_group.options.build(product_option_params)
-        product_option.save!
+        product_option = ProductOption.find_by(channel_code: option[:id])
+        if product_option
+          product_option.update!(name: product_option[:title])
+        else
+          product_option_group.options.build(name: product_option[:title],
+                                             channel_id: channel_id,
+                                             channel_code: option[:id])
+        end
       end
     end
 
+    # == 데이터 포장
     def _make_product_data
       {
         :brand_id => @brand.id,
@@ -74,6 +78,7 @@ module HaravanProduct
       }
     end
 
+    # == 데이터 포장
     def _make_product_option_data(product_option, channel_id)
       {
         name: product_option[:title],
